+++ shinyswatch/_theme_picker.py
@@ -4,37 +4,65 @@ from htmltools import HTMLDependency
 from shiny import reactive, ui
 from shiny.session import require_active_session
 
-from . import __version__ as shinyswatch_version
-from ._bsw5 import BSW5_THEME_NAME, bsw5_themes
-from ._get_theme_deps import deps_shinyswatch_all, deps_shinyswatch_css_files
+from ._assert import DEPRECATED_PARAM, assert_deprecated
+from ._bsw5 import bsw5_themes, bsw5_version
+from ._get_theme import get_theme
+
+DEPRECATED = DEPRECATED_PARAM()
 
 
-def theme_picker_ui(default: BSW5_THEME_NAME = "superhero") -> ui.TagChild:
+def theme_picker_ui(default: DEPRECATED_PARAM = DEPRECATED) -> ui.TagChild:
     """
     Theme picker - UI
 
-    Add this to your UI to enable the theme picker. This function requires :func:`~shinyswatch.theme_picker_server` to be included in your `server` function.
+    Add this to your UI to enable the theme picker. Adds a
+    :func:`~shiny.ui.input_select` for users to choose a shinyswatch theme. This
+    function requires :func:`~shinyswatch.theme_picker_server` to be included in your
+    `server` function.
 
     Notes
     -----
-    * All simultaneous theme picker users on the same Shiny server will see the same theme. This only is an issue when you are sharing the same Shiny server.
+    * Set the initial theme by providing an initial shinyswatch theme to the `theme`
+      argument of a Shiny page function, e.g. :func:`~shiny.ui.page_sidebar`.
     * Do not include more than one theme picker in your app.
     * Do not call the theme picker UI / server inside a module.
+
+    Examples
+    --------
+
+    ```{.python}
+    app_ui = ui.page_sidebar(
+        ui.sidebar(
+            shinyswatch.theme_picker_ui(),
+            # Other input controls in the sidebar...
+        ),
+        # Your main app content...
+        title="Shiny Sidebar Page",
+        theme=shinyswatch.theme.minty, # Initial (optional) app theme
+    )
+    ```
 
     Parameters
     ----------
     default
-        The default theme to be selected when the theme picker is first loaded.
+        Deprecated. Please use the `theme` argument of a Shiny page function to set the
+        initial theme.
 
     Returns
     -------
     :
-        A UI elements creating the theme picker.
+        A UI element creating the theme picker.
 
     See Also
     --------
     * :func:`shinyswatch.theme_picker_server`
     """
+    assert_deprecated(
+        default,
+        "default",
+        "Please use the `theme` argument of a Shiny page function to set the initial theme.",
+    )
+
     return ui.tags.div(
         # Have a div that is hidden by default and is shown if the server does not
         # disable it. This is nice as the warning will be displayed if the server method
@@ -47,20 +75,22 @@ def theme_picker_ui(default: BSW5_THEME_NAME = "superhero") -> ui.TagChild:
         ui.input_select(
             id="shinyswatch_theme_picker",
             label="Select a theme:",
-            # TODO-barret; selected
-            selected=default,
-            choices=bsw5_themes,
+            # These choices are filled in by the server logic
+            selected=None,
+            choices=[],
         ),
-        theme_picker_deps(default),
+        ui.output_ui("shinyswatch_theme_picker_output"),
+        theme_picker_deps(),
     )
 
 
-def theme_picker_deps(initial: str = "superhero") -> list[HTMLDependency]:
+def theme_picker_deps() -> list[HTMLDependency]:
     return [
-        *deps_shinyswatch_all(initial),
         HTMLDependency(
             name="shinyswatch-theme-picker",
-            version=shinyswatch_version,
+            # Uses Bootstrap version so that its served path is the same as the
+            # shinyswatch bootswatch.min.css dependency.
+            version=bsw5_version,
             source={"package": "shinyswatch", "subdir": "picker"},
             stylesheet={"href": "theme_picker.css"},
             script={"src": "theme_picker.js"},
@@ -72,9 +102,19 @@ def theme_picker_server() -> None:
     """
     Theme picker - Server
 
-    This function adds the necessary hooks for the theme picker UI to properly update. This function requires :func:`~shinyswatch.theme_picker_ui` to be included in your UI definition.
+    This function adds the necessary server logic for the theme picker UI to properly
+    update. Be sure to include :func:`~shinyswatch.theme_picker_ui` in your UI
+    definition.
 
-    Note: All simultaneous theme picker users on the same Shiny server will see the same theme. This only is an issue when you are sharing the same Shiny server.
+    Examples
+    --------
+
+    ```{.python}
+    def server(input):
+        shinyswatch.theme_picker_server()
+
+        # The rest of your server logic...
+    ```
 
     See Also
     --------
@@ -85,14 +125,41 @@ def theme_picker_server() -> None:
     input = session.input
 
     @reactive.effect
+    @reactive.event(input.__shinyswatch_initial_theme)
+    def _():
+        choices = set([*ui.Theme.available_presets(), *bsw5_themes])
+        choices = sorted([str(x) for x in choices])
+        if input.__shinyswatch_initial_theme() not in choices:
+            choices = [input.__shinyswatch_initial_theme(), *choices]
+
+        ui.update_select(
+            "shinyswatch_theme_picker",
+            choices=choices,
+            selected=input.__shinyswatch_initial_theme(),
+        )
+
+    @reactive.effect
     @reactive.event(input.shinyswatch_theme_picker)
     async def _():
+        theme = None
+        dep = None
+
+        if input.shinyswatch_theme_picker() in bsw5_themes:
+            theme = get_theme(input.shinyswatch_theme_picker())
+        elif input.shinyswatch_theme_picker() in ui.Theme.available_presets():
+            theme = ui.Theme(preset=input.shinyswatch_theme_picker())
+
+        if theme is not None:
+            dep = theme._html_dependency()  # pyright: ignore[reportPrivateUsage]
+            dep = session._process_ui(dep)[  # pyright: ignore[reportPrivateUsage]
+                "deps"
+            ][0]
 
         await session.send_custom_message(
             "shinyswatch-pick-theme",
             {
                 "theme": input.shinyswatch_theme_picker(),
-                "sheets": deps_shinyswatch_css_files,
+                "dep": dep,
             },
         )