+++ tests/test_themes.py
@@ -1,4 +1,5 @@
 import htmltools
+import pytest
 
 import shinyswatch
 import shinyswatch._bsw5 as bsw5
@@ -16,21 +17,22 @@ def test_error_messages():
 def test_themes():
     for theme_name in bsw5.bsw5_themes:
         theme_obj = shinyswatch.get_theme(theme_name)
-        theme_deps = theme_obj()
+
+        # Assert base theme can use precompiled CSS. If this fails, it's likely that
+        # `py-shinyswatch` is no longer up-to-date with `py-shiny`.
+        cant_precompile = (
+            "Is py-shinyswatch up-to-date with py-shiny? "
+            + "If not, update py-shiny and run `python scripts/update_themes.py`."
+        )
+        assert theme_obj._can_use_precompiled(), cant_precompile
+
+        theme_deps = theme_obj._html_dependency()
 
         # assert all returned html deps are HTMLDependencies
-        assert isinstance(theme_deps, list)
-        for dep in theme_deps:
-            assert isinstance(dep, htmltools.HTMLDependency)
+        assert isinstance(theme_deps, htmltools.HTMLDependency)
+        assert theme_deps.name == f"shinyswatch-css-{theme_name}"
+        assert theme_deps.stylesheet[0]["href"] == "bootswatch.min.css"
 
-        # assert all tagified theme values are HTMLDependencies
-        tagified_theme = theme_obj.tagify()
-        assert isinstance(tagified_theme, htmltools.TagList)
-        for item in tagified_theme:
-            assert isinstance(item, htmltools.HTMLDependency)
-
-        # assert all theme_deps are HTMLDependencies
-        theme_deps = shinyswatch.get_theme_deps(theme_name)
-        assert isinstance(theme_deps, list)
-        for dep in theme_deps:
-            assert isinstance(dep, htmltools.HTMLDependency)
+        # themes are not tagifiable, need to be provided to `ui.page_*(theme=theme_obj)`
+        with pytest.raises(SyntaxError):
+            theme_obj.tagify()